requirements.txt
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
yfinance==0.2.37
python-dotenv==1.0.1
azure-functions==1.21.3
azure-storage-blob==12.25.0